backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
alembic==1.13.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
python-multipart==0.0.9
boto3==1.35.36
slowapi==0.1.9
Jinja2==3.1.4
firebase-admin==6.5.0
structlog==24.1.0
python-json-logger==2.0.7
sentry-sdk[fastapi]==2.15.0
prometheus-fastapi-instrumentator==7.0.0

# Image processing
Pillow==10.4.0  # Image compression

# dev/test
pytest==8.3.3
httpx==0.27.2
pytest-cov==5.0.0
pytest-asyncio==0.23.8

# CLI admin tool
typer==0.12.3
rich==13.7.0
